Function e2ee::ffi::e2ee_server_free

#[no_mangle]
pub unsafe extern "C" fn e2ee_server_free(e2ee_server: *mut E2ee)

Frees the memory associated with an E2ee server object.

This function should be called to release the memory of an E2ee server object that was previously created using Rust code and returned to C. After calling this function, the pointer to the E2ee server object becomes invalid and must not be used again.

Safety

The e2ee_server pointer must be valid and non-null. This function performs an unsafe operation by dereferencing the pointer and taking ownership of the memory, which is then deallocated using Rust’s Box::from_raw().

This function should only be called once for each e2ee_server object. Calling this function multiple times with the same pointer will lead to undefined behavior, as the memory will have already been freed.

The caller must ensure that the pointer was originally allocated by Rust using Box::into_raw(). Passing any other pointer (e.g., a pointer allocated by C or other means) to this function may result in undefined behavior.

Parameters

e2ee_server: A pointer to an E2ee server object (*mut E2ee).

Example

// In C code, after using the server, call this to free its memory:
e2ee_server_free(e2ee_server);

Notes

After this function is called, the memory associated with e2ee_server is freed. The caller should not attempt to access the e2ee_server pointer afterward, as doing so will result in a use-after-free error.