Function e2ee::ffi::e2ee_server_free_string

#[no_mangle]
pub unsafe extern "C" fn e2ee_server_free_string(s: *mut c_char)

Frees the memory associated with a C string.

This function is used to free the memory of a string that was originally allocated by the Rust code and returned to the C code. After calling this function, the pointer to the C string becomes invalid and should no longer be used.

Safety

The s pointer must be valid and non-null. This function performs an unsafe operation by dereferencing the pointer and transferring ownership of the memory to the Rust side. After the function is called, the memory is deallocated. Using the s pointer after this call is undefined behavior.

The function assumes that the pointer was originally created by Rust using CString::into_raw(). Passing any other pointer may result in undefined behavior.

Example

// In C code, after using the string, call this to free it:
e2ee_server_free_string(encrypted_message);

Arguments

s - A pointer to a C string (null-terminated c_char).

Panics

This function may panic if the pointer is invalid or the memory was not allocated by Rust’s CString.