Struct e2ee::server::E2ee

pub struct E2ee { /* private fields */ }

A struct representing the End-to-End Encryption (E2EE) system on the server side.

This struct manages the full lifecycle of key management and encryption operations. It includes:

Key Generation: Generates a key pair (public and private keys) for encryption and decryption.

Encryption: Encrypts messages using the server’s public key.

Decryption: Decrypts messages using the server’s private key.

The server side E2EE system requires both a private key and a public key to function correctly. The private key is used for decryption, while the public key is used for encryption.

The E2ee struct includes the following fields:

private_key: The RSA private key used for decrypting messages.

public_key: The RSA public key used for encrypting messages.

private_key_pem: The PEM-encoded private key as a string.

public_key_pem: The PEM-encoded public key as a string.

Examples

use e2ee::server::{E2ee, KeySize};

// Create a new E2ee instance with a specific key size.
let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");

// Get the PEM-encoded public key.
let public_key_pem = e2ee.get_public_key_pem();
println!("Public Key PEM: {}", public_key_pem);

// Encrypt a message.
let encrypted_message = e2ee.encrypt("Secret message").expect("Failed to encrypt message");
println!("Encrypted Message: {}", encrypted_message);

// Decrypt a message.
let decrypted_message = e2ee.decrypt(&encrypted_message).expect("Failed to decrypt message");
println!("Decrypted Message: {}", decrypted_message);

Errors

The struct’s methods may return errors if key generation fails, or if encryption/decryption operations fail.

Implementations

impl E2ee

pub fn new(key_size: KeySize) -> Result<Self, E2eeError>

Creates a new E2ee instance with the specified key size.

Arguments

key_size - The size of the RSA keys to generate. This should be one of the KeySize variants.

Examples

use e2ee::server::{E2ee, KeySize};

let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");

Errors

This function returns an error if key generation fails.

pub fn new_from_pem( private_key_pem: String, public_key_pem: String, ) -> Result<Self, E2eeError>

Creates a new E2ee instance from PEM-encoded private and public keys.

Arguments

private_key_pem - The PEM-encoded private key as a string.

public_key_pem - The PEM-encoded public key as a string.

Examples

use e2ee::server::E2ee;

let private_key_pem = include_str!("../files/private.pem");
let public_key_pem = include_str!("../files/public.pem");
let e2ee = E2ee::new_from_pem(private_key_pem.to_string(), public_key_pem.to_string())
    .expect("Failed to create E2ee instance from PEM");

Errors

This function returns an error if decoding the PEM keys fails.

pub fn get_public_key(&self) -> &RsaPublicKey

Retrieves the public key in its original RsaPublicKey format.

Examples

use e2ee::server::{E2ee, KeySize};

let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");
let public_key = e2ee.get_public_key();
// Now you can use the public key for encryption

Returns

This function returns a reference to the RsaPublicKey contained in the E2ee struct.

pub fn get_private_key(&self) -> &RsaPrivateKey

Retrieves the public key in its original RsaPrivateKey format.

Examples

use e2ee::server::{E2ee, KeySize};

let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");
let public_key = e2ee.get_private_key();
// Now you can use the public key for encryption

Returns

This function returns a reference to the RsaPrivateKey contained in the E2ee struct.

pub fn get_private_key_pem(&self) -> &str

Retrieves the PEM-encoded private key.

Examples

use e2ee::server::{E2ee,KeySize};

let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");
let private_key_pem = e2ee.get_private_key_pem();
println!("Private Key PEM: {}", private_key_pem);

Returns

This function returns a string slice containing the PEM-encoded private key.

pub fn get_public_key_pem(&self) -> &str

Retrieves the PEM-encoded public key.

Examples

use e2ee::server::{E2ee,KeySize};

let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");
let public_key_pem = e2ee.get_public_key_pem();
println!("Public Key PEM: {}", public_key_pem);

Returns

This function returns a string slice containing the PEM-encoded public key.

pub fn encrypt(&self, message: &str) -> Result<String, E2eeError>

Encrypts a message using the public key.

Arguments

message - The plaintext message to encrypt.

Examples

use e2ee::server::{E2ee, KeySize};

let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");
let message = "Hello, world!";
let encrypted = e2ee.encrypt(message).expect("Failed to encrypt message");

Errors

This function returns an error if encryption fails.

pub fn decrypt(&self, ciphertext: &str) -> Result<String, E2eeError>

Decrypts a ciphertext using the private key.

Arguments

ciphertext - The base64-encoded encrypted message to decrypt.

Examples

use e2ee::server::{E2ee, KeySize};

let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");
let message = "Hello, world!";
let encrypted = e2ee.encrypt(message).expect("Failed to encrypt message");
let decrypted = e2ee.decrypt(&encrypted).expect("Failed to decrypt message");
assert_eq!(message, decrypted);

Errors

This function returns an error if decryption fails.

pub fn save_keys_to_files( &self, private_key_file_path: &str, public_key_file_path: &str, ) -> Result<(), E2eeError>

Saves the PEM-encoded private and public keys to files.

Arguments

private_key_file - The path to the file where the private key PEM should be saved.

public_key_file - The path to the file where the public key PEM should be saved.

Examples

use e2ee::server::{E2ee, KeySize};

let private_key_file_path = concat!(env!("CARGO_MANIFEST_DIR"), "/files/private_key.pem");
let public_key_file_path = concat!(env!("CARGO_MANIFEST_DIR"), "/files/public_key.pem");
let e2ee = E2ee::new(KeySize::Bit2048).expect("Failed to create E2ee instance");
e2ee.save_keys_to_files(private_key_file_path, public_key_file_path)
    .expect("Failed to save keys to files");

// Clean up files
std::fs::remove_file(private_key_file_path)
    .expect("Failed to delete private key file");
std::fs::remove_file(public_key_file_path)
    .expect("Failed to delete public key file");

Errors

This function returns an error if writing to the files fails.